fix(prover-ray): fix the shared randomness contribution
